+++ src/file/file_module.h
@@ -28,9 +28,9 @@ typedef struct _file_header
     int vt_size;
 } header;
 
-void write_file(header *file_header, float *graph, char *filename);
+void write_file(header *file_header, unsigned short int *graph, char *filename);
 header *read_header(char *filename);
-float *read_graph(char *filename, int graph_size);
+unsigned short int *read_graph(char *filename, int graph_size);
 void write_mst(int *v_t, int vt_size, int graph_root, char *filename);
 int *read_mst(char *filename);
 void write_result(char *graph_name, int graph_size, double elapsed_time, int edge_percentage, int num_threads);
